Fix ethereum broadcast issues by using multiple servers
We do not yet simultaneously broadcast to multiple servers because Etherscan will accept a too large a nonce without giving us an error. This will cause numerous problems in the future and cause transactions that never show in the mempool. Blockcypher will return an error allowing us to decrement the nonce and try again. Therefore, broadcast first to Blockcypher then to Etherscan only if Blockcypher errors due to network issues.

* Add logic to broadcast to Blockcypher API
* Add logic to notice transaction validation error due to nonce too high. Only blockcypher returns this.
* Add logic to fallback to Etherscan only if Blockcypher throws an error.
* Refactor nonce logic to increment or decrement the nonce if any server mentions that it needs to
* Add yark.lock file
* Remove package-lock.json