Simplify FIO request fetching
fetchFioRequests was always passing two different arrays to fioRequestsListChanged which was guaranteed to return true. Therefore the number of requests would go back and forth between two values causing inconsistent behavior in the gui. This simplification compares the previously saved array directly to the one queried from the blockchain and saves if it actually changed.

This change in logic means we don't need to pay attention to any in-progress (selected and in the process of paying) requests in fioRequestsToApprove. Since querying the blockchain will always be slightly behind user actions in the gui we can remove those requests from the gui (separate PR) once it's successfully paid instead of removing them from the plugin payload.