Replace checkAccountInnerLoop
Instead we'll listen to events from the Zcash SDK to determine when to retrieve data. The update emitter providers scan and download status and fires rapidly when scanning but quiets down once synced. These events will trigger balance and transaction queries into the SDK. The synchronizer doesn't like when you query transactions or balance when it isn't in synced status so we're watching that too and exiting early from these query functions if it isn't ready. Not perfect, and could probably be better handled in RNZcash.