Move walletConnectors map into ethereum engine global
Any wallet can initiate the websocket connection but the implementation needs to be flexible so the user can choose the wallet to interact with after the connection is opened. That requires moving the connector map to global scope and adapting the calls to accept and pass the walletId so the GUI can use the appropriate wallet to complete contract calls.

(cherry picked from commit dbf07eebe064bdc5c6675979de481b9298c46637)