Add makeSpend and related helpers calcTxFee and txBuilder
To use the Tronscan client we need to overload two of it's methods. First, the addRef() calls getLatestBlock() which is a network call so we override that and provide a reference to an up-to-date reference block for the client to use. Next, we need to add a feeLimit to all TRC20 transactions and since addRef is called for all transactions as can add it there.