Remove multicastServers
XRP uses a single websocket for all wallets so multiServers isn't necessary and required us to recreate types using cleaners. Cleaners are still useful for transactions where there are many, many different types and we just want to drill down into the type we care about. I did make that cleaner a bit more forgiving since some of the fields we enforce are not always present on other types of transactions a user may have completed from another wallet.

The makeSpend change is temporary - we can remove network connectivity from this function in the next commit