Move makeSpend's ERC20 data encoding from the "memoized" gas calculation so it's encoded every time
This is a few bugs. It didn't really matter to users, however, because signTx gets this right and covered up the mistakes in makeSpend. We need to encode the data (correctly) every time we call makeSpend to get an accurate fee especially as we start totaling fees from multiple layers.