Add L1 fee calculation
Optimism transactions need to pay an additional fee that covers the cost of the transaction being rolled up in a batch transaction on the layer 1 chain. There are fixed params, like the raw data overhead and cost per byte, and some others that move according to the underlying chain's fee market.