Add OBT data query
It's very similar to the requests loop but within enough differences that making generic was too complicated. I'm also attaching these to the engine instead of writing to otherData because I don't understand why any of this queried data is being put on disk anyway so I'm not going to add to it. Only the highestTxHeight should remain in otherData AFAICT.