Remove fioRequestsToApprove
The purpose of why this function was created is a mystery. My best guess is the engine was unreliable in the past and it was desirable to spam the network with the recordObtData transactions until they worked. This function was created before the engine dynamically changed the local sdk instance baseUrl so I suspect the engine would start with a bad baseUrl and fail to submit until the user rolled the dice with a new baseUrl after relaunching the app and this loop tried it again.

The transaction flow has changed and the problem I described hasn't existed for some time since we know use all urls to broadcast transactions. If we wanted to keep the redundancy it becomes difficult (not impossible) to keep this using the new flow and is only briefly useful. We could store the signedTx object instead of params and try rebroadcasting but it's only useful to do so until the signedTx expiration.

tldr: I don't think it's necessary so I removed it.