Move FIO get names response cleaning to calling function
promiseNy shouldn't be cleaning the response and only cares about the return value as-is to compare to others. The cleaning can happen after multiple responses are equal. 

It's not possible to pass the cleaner before the account has activity so every response would just be undefined and the function call would pollute the logs for no reason.