Fix missing last page query in FIO history sync
The last page in the history query is the first transactions for the
account. This page was skipped because the `pos` could go negative
causing an early break out of the loop. The fix is to limit the `pos`
to zero, and let the final query finish before checking whether to
break out of the loop from one place in the loop body.

This commit also renames `HISTORY_NODE_OFFSET` to
`HISTORY_NODE_PAGE_SIZE` because the constant is a page size constant
and not an offset.