Add autoshielding
We need to overstuff syncNetwork In order to get access to the keys to create an autoshielding transaction. To initiate one, we identify in the balance event that we have excess transparent funds and then force syncNetwork to restart and then create the transaction. Once we see it mined we can release the lock and continue looking for excess transparent funds