Restart the synchronizer if it stops responding
A bug in the synchronizer block downloader logic will cause all activity to stop completely without warning or errors. It just stops downloading and parsing blocks and hangs. Since the synchronizer is noisy at boot we can restart it if we haven't heard from it in while.

It's unknown if the bug occurs after the initial sync is complete and the synchronizer is just listening for the latest blocks. I'm assuming it does not and am limiting this method to just initial sync.