Simplify NetworkAdapter interface by making it an abstract class
The `NetworkAdapterBase` becomes an abstract class, `NetworkAdapter`,
and extended classes can implement this abstract class with all the
utilities included with the class along with the documentation on the
abstract fields required to be implemented.

This is a lot nicer/easier to work with, within a good IDE like VSCode.
It also makes it simpler because now we can avoid the
`extends...implements` with just a single `extends`.