Ignore strange denoms with hyphens that the sdk utils can't parse
Hyphens don't appears in the regex tested by parseCoins so weird denoms on chains like Coreum were failing. There doesn't appear to be official documentation on denom being different on Coreum vs Cosmos so we're just going to ignore them for now.

The only known example of a hyphenated denom is an airdrop token unique to a holders address.