Add common types to registry
By default, a registry only know a couple of send transaction types. We'll need to know a couple more so that we can parse them and show important information about them. We won't be able to parse everything (there's a ton) but these two appeared in testing so they've been added