Remove `LastEstimatedGasLimit` caching mechanism
The reason for this caching appears to only be an optimization. This
optimization cannot be used because zkSync's gas estimation changes when
the transaction value parameter changes. Because chains like zkSync may
change the gas estimation based on the parameters, there is no reason to
cache unless all the parameters are the same. However, there is no
UX instance where the estimation call is made with identical parameters
to warrant the optimization.

Instead, anywhere that the caching is needed, whether for optimization
or for correctness, the implementation for the caching should be done
within that routine.