Add ADR-036 support
This is an amino message with special rules used for wallets to be able to sign messages and not transactions. We don't want to bother parsing amount information from it, because there is none, and we need to verify that it meets all the rules. Valid and non-ADR036 messages will pass the validation function.