Enforce the network address minimum on SOL max spends
This is an unfortunate change but one that's necessary to support wallet sweeping which creates multiple transactions rapidly without waiting for confirmation. The bug in that case is the node evaluates the max SOL send (which has accounted for the fees included in the token sends) but since there isn't a nonce the node doesn't consider the balance changes in the token transaction before evaluating the SOL. The SOL tx then appears to have too little of a balance remaining and the tx is rejected

This might be resolvable by integrating nonce accounts or maybe we can stuff all the token transactions and final SOL spend into a single transaction.