Include accountKey and privateKey in CardanoPrivateKey specification
The `accountKey` is the key on path `m/1852'/1815'/0'` which takes
significant computation time to derive because it is multiple levels
of harden derivations. Usually, this key is stored as formatted key
in plugins like the UTXO plugins as an optimization.

The `privateKey` is the root key (`m/`) which also takes significant
computation type to derive from mnemonic. This root-key isn't really
used anywhere except for deriving the `publicKey`, which returned by
`getDisplayPublicKey`.

This change includes this optimization in a backwards compatible way.
This means newer Cardano wallets will benefit from the optimization
only.