Add staking methods to otherMethods for CardanoEngine
The method to get stake address is specifically a Cardano concept, so
it makes sense to keep it as an otherMethods function rather than
putting the stake address in the standard get address methods.

The decode function includes validating the transaction is a staking
transaction which means that all outputs are the wallet's own outputs
and it includes certain certificates for staking delegation and a
withdrawals field.