Disable 'blockbook-ws' connections
These are preventing incoming transactions from appearing.

Because these adapters exist in the list, we call `setupAdapterSubscriptions` on them, which prevents the polling loop from checking any addresses.

Unfortunately, the two WebSocket URL's don't actually work, so we never actually subscribe to the addresses on the WebSocket side either.

This means that updates never appear, so the only way to see incoming transactions is to resync and log out.

This is a temporary solution to solve the immediate problem. The long-term solution is to fix the engine to poll for addresses unless there are active, live, working-for-real WebSocket subscriptions.