Optional: Break up transaction query to provide progress updates
By far, the slowest query is `address_txs` and we don't know how many times we have to query it but it does return 1000 txs at a time. We can't provide any progress updates until we get the list of txids and break them up.

The first query takes so long that the end result isn't remarkably better than the original. Breaking up the second info query into chunks technically makes syncing take longer with the "benefit" that maybe we'll see actual progress in the sync ratio. This commit is optional and I can be convinced it isn't necessary.