Fix UTXO stomping from the network for Cardano
Network queries would result in overwriting local UTXO state without
considering that the UTXOs may have already been spent and the network
is out of date.

The solution is to always merge unseen UTXOs from the network into the
the wallet state, and ignore UTXOs that have already been seen by the
wallet.

This algorithm assumes that the network will catch-up with the wallet
well before the wallet is restarted/resync'd because the `seenVouts`
state is ephemeral. If it were to persist engine restarts, then the
only bug surface area is if the user resyncs (destroying any
persisted state for the wallet). This is generally how UTXO plugins work,
but doesn't appear to be necessary for Cardano due to it's fast block
times.