Use Helius RPC archival node exclusively
The existing archival nodes were failing or too slow to sync. Using
Helius RPC node exclusively allows us to sync transactions quickly.

A future patch would be to fix the issues with the other two archival
nodes. Known issues with `solana.com` node is a rate limiting issue.
The issue with `alchemy.com` seems to be wide spread. The RPC node
seems to be returning responses that are unexpected to the web3 Solana
library. More debugging is needed to get these services back to a
working state.