Fix new transaction notifications
We had a cacth-22 situation, where we needed a valid checkpoint to consider any transaction new, but we wouldn't update the checkpoint until we saw a new transaction.

Besides the catch-22, a fully-syncd wallet generally wouldn't have any activity that could update the checkpoint, leaving it stuck at undefined until a new transaction comes in, in which case that new transaction would never get a notification. Therefore, we need to set up an initial checkpoint at boot if we don't have one. This will be a one-time operation on a single device. After that, new devices will recieve the synced checkpoint, and incoming transactions will receive notifications.