Support "silent sync" enabled by change-server
The wallet will silently sync the wallet without any network request
if the change-server subscription event contains a checkpoint that is
behind the local checkpoint (block-height) of the engine.
This assumes the core will always call an initial `syncNetwork` if the
change-server responds with a successful subscription without any
changes needed.