Add getTokenDetails to Solana
Solana tokens don't have native metadata so we're using another program, Metaplex, to find metadata. It a program that token minters opt into to so it's not available for every token but it seems to be popular and have decent coverage. There may be additional programs or 3rd party tools that track metadata available but this is a decent start.