Fix thrashing in syncNetwork
By catching unknown errors in the acquireUpdates function, the caller
doesn't know that the function failed and will retry after snoozing.
The wallet would be stuck in this loop thrashing the network and
degrading device performance while the app is in a buggy state because
of the unknown error.
Instead, we'll break out of the syncNetwork function because the unknown
error is actually thrown. The core should log errors from syncNetwork
to be introspected by a developer when debugging.