Fix syncNetwork sync accuracy
The first problem solved with this change is to have a dedicated
`highestTxBlockHeight` state which tracks the blockHeight of the last
transaction received. This would more accurately determine whether a
change is present on the network and eliminate missing transaction bugs.

The second fix is to increase the immediacy of the call to
`onSubscribeAddresses` callback. This gets the subscription request to
the core sooner, minimizing the race for the core to start the
`PeriodicTask` version of `syncNetwork`.