Disable connection based network adapters
This disables the BlockbookWsAdapter which is the only connection based
adapter currently.

The reason for this change is because the blockbook server connections
are not working due to rate limiting and after the new architecture
with the change server, these connection based adapters don't really
improve performance or function properly.