Add support for multiple subscan urls
The history of polkadot is now split between two chains so we need to use two subscan urls to get the transaction history. For the old chain, we can mark those transactions as confirmed since they can cannot be compared against the block height of the new chain. For simplicity, we only count the transactions on the latest chain towards the sync ratio.