Don't set FIO tx blockheight in broadcastTx
There's an obvious timing issue here where the transaction could be assigned a blockheight we're not aware of. It's a transaction from the future until checkBlockchainInnerLoop runs again. But on deeper inspection, the height returned in the node's response is not actually the confirmation height it's just the network height when the transaction was submitted. That means the transaction hasn't fully confirmed and therefore the field should not be edited until the transaction loop sees it.